Script Clean-Up & Bug Fix + Alt controls + Archer
knock back script passes along variable to revert back to vector2.zero and idle state rather than invoking a call in the knockback script.  This is import because it prevents a situation where the gameobject can be destroyed before the call back to vector2.zero & idle.

Now keyboard controls are working

archer model is available for player.